requirements.txt
pytest
pytest-check
pytest-expect
delayed-assert
allure-pytest
requests
beautifulsoup4
bs4
selenium
lxml
parse
psutil
urllib3
pytest-subtests
webdriver-auto-update
urllib3
opencv-python
matplotlib
requests-toolbelt
pandas
numpy
Appium-Python-Client
pytz
webdriver_manager